Rename header to standard one
